Split rocjpeg code (#3)
* Update decode_jpegs_cuda.cpp

* Update decode_jpegs_cuda.h

* split code
